pyproject.toml
[project]
name = "openllm-agent-synth"
version = "0.1.0"
description = "A framework for generating datasets for agentic applications."
readme = "README.md"
authors = [
  {name = "Loong Ma", email = "diqiuzhuanzhuan@gmail.com"}
]
maintainers = [
  {name = "Loong Ma", email = "diqiuzhuanzhuan@gmail.com"}
]
classifiers = [
    # TODO
]
license = {text = "MIT"}
dependencies = [
  "data-designer>=0.5.6",
  "pydantic",
  "python-dotenv",
  "PyYAML",
  "rich",
  "typer",
]
requires-python = ">= 3.10"

[project.optional-dependencies]
test = [
    "coverage",  # testing
    "pytest",  # testing
    "ruff",  # linting
    "ty", # checking types
    "ipdb", # debugging
]

[project.urls]
bugs = "https://github.com/diqiuzhuanzhuan/openllm-agent-synth/issues"
changelog = "https://github.com/diqiuzhuanzhuan/openllm-agent-synth/blob/master/HISTORY.md"
homepage = "https://github.com/diqiuzhuanzhuan/openllm-agent-synth"

[project.scripts]
openllm-agent-synth = "openllm_agent_synth.cli:app"

[tool.ty]
# All rules are enabled as "error" by default; no need to specify unless overriding.
# Example override: relax a rule for the entire project (uncomment if needed).
# rules.TY015 = "warn"  # For invalid-argument-type, warn instead of error.

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]

[tool.uv]
package = true
